Rename folder to contractManager
