+++ src/contractManager/ContractV111Utils.ts
