+++ src/contractManager/ContractVersionUtils.ts
